refactor(text-plugin): Remove circulare dependencies
