+++ packages/plugins/text/src/types/config.ts
@@ -2,7 +2,7 @@ import { SerializedScalarStateType } from '@edtr-io/plugin'
 import { DeepPartial } from '@edtr-io/ui'
 import { Descendant, Range } from 'slate'
 
-import { Heading } from '.'
+import { Heading } from './text-editor'
 
 /** @public */
 export type TextEditorState = SerializedScalarStateType<
